test: union
